tag component changes
